+++ sdk/cosmos/azure-cosmos/azure/cosmos/_constants.py
@@ -94,5 +94,6 @@ class _Constants:
 
         RETRY_WRITE: Literal["retry_write"] = "retry_write"
         """Whether to retry write operations if they fail. Used either at client level or request level."""
+        EXCLUDED_LOCATIONS: Literal["excludedLocations"] = "excludedLocations"
         AVAILABILITY_STRATEGY: Literal["availability_strategy"] = "availability_strategy"
         """Availability strategy. Used either at client level or request level"""